+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased]
 
+### Changed
+- [harness] Package and workspace directory renamed from fleet-harness-extension to fleet-harness
+
 ## [0.16.0] - 2026-05-10
 
 ### Added